Create an AddOnlyOrderedMap instance

This function initializes a new list-like object using the specified keys and values if provided. The resulting map preserves the order of insertion and does not allow modification or removal of existing entries.

The AddOnlyOrderedMap can be interacted with exactly like a regular list, possessing methods for [, [[, [<-, and [[<- with the same behaviour, except that NULL cannot be passed in since removal is not permitted.

The keys() generic is defined for this class, which will return a list of the keys in their insertion order. The has_key() generic is also defined for this class, returning TRUE/FALSE if a key exists. Lastly, the values() generic is defined to get all values in insertion order.

Usage

AddOnlyOrderedMap(
  keys = NULL,
  values = NULL,
  key_validator = Negate(is.null),
  val_validator = Negate(is.null)
)

Arguments

keys

Optional list. A vector of keys to initialize the map with. Can be any R object. It is assumed that all keys are unique, otherwise the behaviour is undefined. This CANNOT be a scalar value. If that is desired, wrap it in a list().

values

Optional list. An iterable vector of values corresponding to the keys. This CANNOT be a scalar value. If that is desired, wrap it in a list().

key_validator

Optional function. A function that validates individual keys before insertion, returning TRUE if valid, FALSE otherwise.

val_validator

Optional function. A function that validates individual values before insertion, returning TRUE if valid, FALSE otherwise.

Value

An AddOnlyOrderedMap instance

Details

The average time complexity of all operations are linear with respect to the number of insertion/query inputs, in contrast to R lists which has quadratic time complexity for the same operations.

Examples

map <- AddOnlyOrderedMap(keys = c("a", "b"), values = c(1, 2))
#> Error in AddOnlyOrderedMap(keys = c("a", "b"), values = c(1, 2)): could not find function "AddOnlyOrderedMap"

# With validators
key_val <- function(k) is.character(k)
val_val <- function(v) is.numeric(v)
map_validated <- AddOnlyOrderedMap(
  keys = c("a", "b"),
  values = c(1, 2),
  key_validator = key_val,
  val_validator = val_val
)
#> Error in AddOnlyOrderedMap(keys = c("a", "b"), values = c(1, 2), key_validator = key_val,     val_validator = val_val): could not find function "AddOnlyOrderedMap"