Singleton Backend Registry, managing all available backends.

This class should never be instantiated directly, and only used internally via get_backend_registry()

Super class

R6P::Singleton -> BackendRegistry

Methods

Public methods

BackendRegistry$get_backend()

BackendRegistry$set_debug()

BackendRegistry$register_backend()

BackendRegistry$clone()

Inherited methods

R6P::Singleton$initialize()

Method get_backend()

detect the return relevant backend from the input

Usage

BackendRegistry$get_backend(tensor)

Arguments

tensor

any support tensor-like class

Returns

A singleton instance of a BackendRegistry() object

Method set_debug()

Set whether debug messages should be displayed

Usage

BackendRegistry$set_debug(flag = TRUE)

Arguments

flag

boolean

Returns

this object

Method register_backend()

Register a new backend singleton

Usage

BackendRegistry$register_backend(backend)

Arguments

backend

an EinopsBackend subclass

Returns

this object

Method clone()

The objects of this class are cloneable with this method.

Usage

BackendRegistry$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.