Constructor for an Execution Plan (CookedRecipe in the python sourcecode)

Constructor for an Execution Plan (CookedRecipe in the python sourcecode)

Usage

EinopsExecutionPlan(
  init_shapes,
  axes_reordering,
  reduced_axes,
  added_axes,
  final_shapes,
  n_axes_w_added
)

Arguments

init_shapes

integer vector specifying initial tensor shapes for reshaping. Length 0 signifies nullness.

axes_reordering

integer vector specifying the order for transposing tensor axes. Length 0 signifies nullness.

reduced_axes

List of integers specifying which axes to reduce during operations. Length 0 signifies nullness.

added_axes

r2r::hashmap() mapping axis positions (int) to their lengths (int) for axes to be added.

final_shapes

list of integers specifying final tensor shapes for reshaping. Length 0 signifies nullness.

n_axes_w_added

Integer specifying the total number of axes after adding new axes.

Value

An object of class EinopsExecutionPlan, which is a list containing the execution plan for transforming tensors according to the specified recipe.