Allows reordering elements and repeating them in arbitrary combinations.

This operation includes functionality of repeat, tile, and broadcast functions.

Usage

einops.repeat(x, expr, ...)

repeat x

Arguments

x

tensor: array, matrix, or list of arrays of the same shape and type

expr

string: reduction pattern

...

either corresponding axes lengths or a single list of them.

Value

tensor of the same type as input, with dimensions according to output pattern

Details

When composing axes, C-order enumeration is used (consecutive elements have different last axis). Find more examples in the vignettes.

Why can't the function be called as repeat()?

repeat is a reserved keyword in R that acts the same as while(TRUE), and has no way of being overridden. Hence, this function can only be called as einops.repeat() or using backticks as `repeat`().

Examples

# a grayscale image (of shape height x width)
image <- array(rnorm(30*40), dim = c(30, 40))

# change it to RGB format by repeating in each channel
einops.repeat(image, 'h w -> h w c', c=3)
#> Error: In Einops - length(dim(x)) not equal to length(repeats)

# repeat image 2 times along height (vertical axis)
einops.repeat(image, 'h w -> (r h) w', r=2)
#> Error: In Einops - length(dim(x)) not equal to length(repeats)

# repeat image 2 times along height and 3 times along width
einops.repeat(image, 'h w -> (h2 h) (w3 w)', h2=2, w3=3)
#> Error: In Einops - length(dim(x)) not equal to length(repeats)

# convert each pixel to a small square 2x2, i.e. upsample an image by 2x
einops.repeat(image, 'h w -> (h h2) (w w2)', h2=2, w2=2)
#> Error: In Einops - length(dim(x)) not equal to length(repeats)

# 'pixelate' an image first by downsampling by 2x, then upsampling
# downsampled <- reduce(image, '(h h2) (w w2) -> h w', 'mean', h2=2, w2=2)
# einops.repeat(downsampled, 'h w -> (h h2) (w w2)', h2=2, w2=2)